requirements.txt
asgiref==3.8.1
botocore==1.40.36
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudinary==1.44.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-extensions==4.1
django-tailwind==4.0.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
mercadopago==2.3.0
mysqlclient==2.2.7
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
reportlab==4.4.4
requests==2.32.3
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.11.0
